remastered tests
